chore: publish release artifact
GitHub had promised that their source archives would have stable checksums, but then there was a big Bazel outage yesterday due to their changing it.
Even if they do make such a guarantee and stick to it, we also have the problem that GitHub doesn't give download metrics on their source archive so this is still worth changing.