feat(cd): auto-tag releases twice a month
Avoids the common pitfall where users have to request a release, i.e. the inevitable slack message
'hey could you cut a release with that new fix'

Relies on the repository using semver commit messages. This is already implied in this template because
commitizen-tools/commitizen is installed as a commit-msg pre-commit hook.

Based on https://github.com/aspect-build/rules_lint/pull/427/files where this is already observed to be working.