fix: don't set module version outside BCR (#148)
Leaving the version unset or set to the default value ("") can prevent
issues when the module is used via non-registry overrides (e.g.
https://github.com/bazel-contrib/rules_go/issues/4380).

The publish.yaml GitHub Action sets the version in the registry to the
release version by patching the MODULE.bazel file in the pull request to
the BCR.

Also, set compatibility_level back to the default by commenting it out
and explicitly setting it to the default value in the comment and adding
a note about the best practice re. when and how to bump it.

Thanks to fmeum for all the help and context regarding these issues!

For more info, see this Slack thread:
https://bazelbuild.slack.com/archives/CA31HN1T3/p1750406404452179